Refactor components + Responsive scroll
